minor fixes

+++ projektfiler/Voter.ts
@@ -3,8 +3,6 @@ import { GlobalFunctions } from "./GlobalFunctions";
 import Discord from "discord.js";
 import { Database } from "sqlite3";
 
-
-
 export class Voter {
     public static showVotes(message: Discord.Message, args: any[]) {
         if (!args[0]) {
@@ -42,7 +40,11 @@ export class Voter {
 
     private static timedOutUsers: Map<string, number> = new Map(); //this is a cache of users that are timed out in order to ease the load for the database
 
-    static async vote(message: Discord.Message, args: string[]): Promise<void> {
+    static async vote(
+        message: Discord.Message,
+        args: string[],
+        dm?: boolean
+    ): Promise<void> {
         if (!args[0]) {
             message.channel.send(
                 "please use correct input values, !vote [userId] [section]"
@@ -56,7 +58,7 @@ export class Voter {
             section = section.concat(args[i] + " "); //turns section into a string
         }
         section = section.slice(0, -1);
-        message.delete(); //removes the message before the queries in order to delete it as fast as possible
+        if (!dm) message.delete(); //removes the message before the queries in order to delete it as fast as possible
         if (votee == voter.id) {
             voter.send("You can't vote for yourself");
             return;
@@ -157,10 +159,16 @@ export class Voter {
         });
     }
 
-    public static async tallyVotes(client: Discord.Client, message: Discord.Message, args?: string[]): Promise<void> {
+    public static async tallyVotes(
+        client: Discord.Client,
+        message: Discord.Message,
+        args?: string[]
+    ): Promise<void> {
         return new Promise((resolve) => {
             if (GlobalFunctions.messageIsDirectMessage(message)) {
-                message.author.send("this command can not be sent through direct messages. Please type in a channel");
+                message.author.send(
+                    "this command can not be sent through direct messages. Please type in a channel"
+                );
                 resolve();
                 return;
             }
@@ -172,14 +180,15 @@ export class Voter {
             section = section.slice(0, -1);
             console.log(section.length + " " + section);
             if (section.length == 0)
-                query = "SELECT section, votee, COUNT(votee) as votes FROM Votes GROUP BY section, votee;";
+                query =
+                    "SELECT section, votee, COUNT(votee) as votes FROM Votes GROUP BY section, votee;";
             else
-                query = "SELECT section, votee, COUNT(votee) as votes FROM Votes WHERE section = ? GROUP BY section, votee;";
+                query =
+                    "SELECT section, votee, COUNT(votee) as votes FROM Votes WHERE section = ? GROUP BY section, votee;";
             let db = DatabaseFunctions.getInstance();
             let runnable = db.prepare(query);
             let handler = async (err, res) => {
-                if (err)
-                    console.log(err);
+                if (err) console.log(err);
                 if (!res) return;
                 this.sortQuery(res).then((result) => {
                     let iterator = result.entries();
@@ -191,65 +200,100 @@ export class Voter {
                     }
                 });
             };
-            if (section.length > 0)
-                runnable.all(section, handler);
-            else
-                runnable.all(handler);
+            if (section.length > 0) runnable.all(section, handler);
+            else runnable.all(handler);
             resolve();
-        })
+        });
     }
 
     private static async delay(ms: number) {
-        return new Promise(resolve => setTimeout(resolve, ms));
+        return new Promise((resolve) => setTimeout(resolve, ms));
     }
 
-
-    private static async addSectionToEmbed(client: Discord.Client, value: [section: string, rows: { section: string, votee: string, votes: number }[]], message: Discord.Message): Promise<void> {
+    private static async addSectionToEmbed(
+        client: Discord.Client,
+        value: [
+            section: string,
+            rows: { section: string; votee: string; votes: number }[]
+        ],
+        message: Discord.Message
+    ): Promise<void> {
         return new Promise((resolve) => {
-            let embed = new Discord.MessageEmbed().setTitle(value[0]).setColor("#00EF00"); //Created an embed with the title of the section as well as the color green
-            this.addAllSubToEmbed(client, value[1], embed).then(() => { this.delay(2000).then(() => { message.channel.send(embed) }) }) //delay amount is dependant on hardware and has to be hardcoded
-            resolve(); //since the amount of nominees per section is maxed out at 5 we have an upper bound for time at least
-        })
-    }
-
-    private static async addAllSubToEmbed(client: Discord.Client, rows: { section: string, votee: string, votes: number }[], embed: Discord.MessageEmbed): Promise<void> {
-        return new Promise((resolve) => {
-            rows.forEach((row) => { //for each votee
-                GlobalFunctions.idToUsernameClient(client, row.votee).then(user => { //get username from id
-                    embed.addField(user.username, "Votes: " + row.votes, true); //add name and votes in row
-                    console.log("added"); //debugging purposes, should all be after "added subsection"
-                    resolve();
+            let embed = new Discord.MessageEmbed()
+                .setTitle(value[0])
+                .setColor("#00EF00"); //Created an embed with the title of the section as well as the color green
+            this.addAllSubToEmbed(client, value[1], embed).then(() => {
+                this.delay(2000).then(() => {
+                    message.channel.send(embed);
                 });
+            }); //delay amount is dependant on hardware and has to be hardcoded
+            resolve(); //since the amount of nominees per section is maxed out at 5 we have an upper bound for time at least
+        });
+    }
+
+    private static async addAllSubToEmbed(
+        client: Discord.Client,
+        rows: { section: string; votee: string; votes: number }[],
+        embed: Discord.MessageEmbed
+    ): Promise<void> {
+        return new Promise((resolve) => {
+            rows.forEach((row) => {
+                //for each votee
+                GlobalFunctions.idToUsernameClient(client, row.votee).then(
+                    (user) => {
+                        //get username from id
+                        embed.addField(
+                            user.username,
+                            "Votes: " + row.votes,
+                            true
+                        ); //add name and votes in row
+                        console.log("added"); //debugging purposes, should all be after "added subsection"
+                        resolve();
+                    }
+                );
             });
             resolve();
         });
     }
 
-    private static async sortQuery(queryResult: { section: string, votee: string, votes: number }[]): Promise<Map<string, { section: string, votee: string, votes: number }[]>> {
+    private static async sortQuery(
+        queryResult: { section: string; votee: string; votes: number }[]
+    ): Promise<
+        Map<string, { section: string; votee: string; votes: number }[]>
+    > {
         return new Promise((resolve, reject) => {
-            let sections: Map<string, { section: string, votee: string, votes: number }[]>;
+            let sections: Map<
+                string,
+                { section: string; votee: string; votes: number }[]
+            >;
             sections = new Map();
-            queryResult.forEach(row => {
-                if (!sections.has(row.section)) //if row is of new section
+            queryResult.forEach((row) => {
+                if (!sections.has(row.section))
+                    //if row is of new section
                     sections.set(row.section, []); //create new
                 sections.get(row.section).push(row); //add row to set of other rows where section is the same
             });
             let iterator = sections.keys(); //with sorted map of arrays of rows
             let key = iterator.next();
             while (!key.done) {
-                sections.set(key.value, this.sectionSort(sections.get(key.value))); //use section sort to trim it down to the highest 5 voted in each section
+                sections.set(
+                    key.value,
+                    this.sectionSort(sections.get(key.value))
+                ); //use section sort to trim it down to the highest 5 voted in each section
                 key = iterator.next();
             }
             resolve(sections); //return a Map of every section linked to the top 5 votee for each section
         });
     }
 
-    private static sectionSort(section: { section: string, votee: string, votes: number }[]): { section: string, votee: string, votes: number }[] {// take arry of rows sort by votes and give five highest 
-        section.sort((a, b) => a[2] > b[2] ? -1 : 1);
+    private static sectionSort(
+        section: { section: string; votee: string; votes: number }[]
+    ): { section: string; votee: string; votes: number }[] {
+        // take arry of rows sort by votes and give five highest
+        section.sort((a, b) => (a[2] > b[2] ? -1 : 1));
         let result = section.slice(0, 5);
         return result;
     }
-
 }
 
 enum result {
